added Intelli-J to .gitignore; update pom.xml; restructured project - rename directories;
