+++ .gitignore
@@ -1,7 +1,9 @@
+#Intelli-J
 .idea/
 */resources/
 target/
 out/
+
 #Visual Studio
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
