+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: com.vertexinc.returns.test.cloudui.TestMainApplication
+Main-Class: io.github.nezonium.TestMainApplication
