+++ src/main/java/io/github/nezonium/util/PageInterface.java
@@ -1,4 +1,4 @@
-package com.vertexinc.returns.test.cloudui.util;
+package io.github.nezonium.util;
 
 public interface PageInterface {
     /**
